move robot related things to new file
